Restructured models to internal folder
